requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.1
geoip2==2.2.0
gunicorn==19.4.5
ipaddr==2.1.11
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
maxminddb==1.2.0
python-dateutil==2.5.0
python-geoip==1.2
requests==2.9.1
six==1.10.0
SQLAlchemy==1.0.12
Werkzeug==0.11.4
